+++ src/DynAI/Algorithms/AlgorithmTypeEnum.cs
@@ -1,0 +1,14 @@
+﻿using Autodesk.DesignScript.Runtime;
+
+namespace AI.Algorithms
+{
+    [IsVisibleInDynamoLibrary(false)]
+    public enum AlgorithmType
+    {
+        Regression,
+        Classifier,
+        Clustering,
+        DecisionTree,
+        VectorMachine
+    }
+}
